requirements.txt
streamlit
langchain
langchain-openai
langchain-core
python-dotenv
pypdf
python-docx
youtube-transcript-api